% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove_shiny_inputs}
\alias{remove_shiny_inputs}
\title{Remove Shiny Inputs}
\usage{
remove_shiny_inputs(id, .input)
}
\arguments{
\item{id}{character vector of full input ids to remove}

\item{.input}{an input object from a shiny session}
}
\value{

}
\description{
Remove Shiny Inputs
}
