% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-table.R
\name{tidy_table}
\alias{tidy_table}
\title{The NSE table}
\usage{
tidy_table(...)
}
\description{
An additional class to extend \code{data.frame} to allow
for NSE filtering on \code{i}, and tidyselect functionality on \code{j}.
This will help condense some code statements.
}
